Moved all the images into images folder, moved the router and styles into a config folder
