+++ src/features/start/startOrJoinGame.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import {Text, View, StyleSheet, Animated, TextInput, TouchableOpacity} from 'react-native';
-import styles from './styles'
+import styles from './config/styles'
 
 interface Props {
   navigation: any,
